useImageCarousel hook
Moved logic into useImageCarousel hook. Slight style changes.
