+++ src/components/Projects/index.css
@@ -127,10 +127,11 @@
     background: rgba(255, 255, 255, 0.5);
     cursor: pointer;
     transition: background 0.3s ease;
+    background-color: lightgray;
 }
 
 .indicator.active {
-    background: white;
+    background: grey;
 }
 
 .indicator:hover {
